+++ src/main/java/com/braintreegateway/CreditCardVerificationRequest.java
@@ -4,6 +4,8 @@ public class CreditCardVerificationRequest extends Request {
 
     private CreditCardVerificationCreditCardRequest creditCardRequest;
     private CreditCardVerificationOptionsRequest optionsRequest;
+    private ExternalVaultVerificationRequest externalVaultVerificationRequest;
+    private RiskDataVerificationRequest riskDataVerificationRequest;
 
     public CreditCardVerificationRequest() {
     }
@@ -18,6 +20,16 @@ public class CreditCardVerificationRequest extends Request {
         return optionsRequest;
     }
 
+    public ExternalVaultVerificationRequest externalVault() {
+        externalVaultVerificationRequest = new ExternalVaultVerificationRequest(this);
+        return externalVaultVerificationRequest;
+    }
+
+    public RiskDataVerificationRequest riskData() {
+        riskDataVerificationRequest = new RiskDataVerificationRequest(this);
+        return riskDataVerificationRequest;
+    }
+
     @Override
     public String toXML() {
         return buildRequest("verification").toXML();
@@ -26,7 +38,9 @@ public class CreditCardVerificationRequest extends Request {
     protected RequestBuilder buildRequest(String root) {
         RequestBuilder builder = new RequestBuilder(root).
             addElement("creditCard", creditCardRequest).
-            addElement("options", optionsRequest);
+            addElement("options", optionsRequest).
+            addElement("externalVault", externalVaultVerificationRequest).
+            addElement("riskData", riskDataVerificationRequest);
         return builder;
     }
 }