+++ src/test/java/com/braintreegateway/integrationtest/WebhookNotificationIT.java
@@ -528,12 +528,31 @@ public class WebhookNotificationIT extends IntegrationTest {
     }
 
     @Test
-    public void createsSampleNotificationForGrantedPaymentInstrumentUpdate() {
-        HashMap<String, String> sampleNotification = this.gateway.webhookTesting().sampleNotification(WebhookNotification.Kind.GRANTED_PAYMENT_INSTRUMENT_UPDATE, "my_id");
+    public void createsSampleNotificationForGrantorUpdatedGrantedPaymentMethod() {
+        HashMap<String, String> sampleNotification = this.gateway.webhookTesting().sampleNotification(WebhookNotification.Kind.GRANTOR_UPDATED_GRANTED_PAYMENT_METHOD, "my_id");
 
         WebhookNotification notification = this.gateway.webhookNotification().parse(sampleNotification.get("bt_signature"), sampleNotification.get("bt_payload"));
 
-        assertEquals(WebhookNotification.Kind.GRANTED_PAYMENT_INSTRUMENT_UPDATE, notification.getKind());
+        assertEquals(WebhookNotification.Kind.GRANTOR_UPDATED_GRANTED_PAYMENT_METHOD, notification.getKind());
+
+        GrantedPaymentInstrumentUpdate update = notification.getGrantedPaymentInstrumentUpdate();
+
+        assertEquals("vczo7jqrpwrsi2px", update.getGrantOwnerMerchantId());
+        assertEquals("cf0i8wgarszuy6hc", update.getGrantRecipientMerchantId());
+        assertEquals("ee257d98-de40-47e8-96b3-a6954ea7a9a4", update.getPaymentMethodNonce());
+        assertEquals("abc123z", update.getToken());
+        assertEquals("expiration-month", update.getUpdatedFields().get(0));
+        assertEquals("expiration-year", update.getUpdatedFields().get(1));
+        assertEquals(2, update.getUpdatedFields().size());
+    }
+
+    @Test
+    public void createsSampleNotificationForRecipientUpdatedGrantedPaymentMethod() {
+        HashMap<String, String> sampleNotification = this.gateway.webhookTesting().sampleNotification(WebhookNotification.Kind.RECIPIENT_UPDATED_GRANTED_PAYMENT_METHOD, "my_id");
+
+        WebhookNotification notification = this.gateway.webhookNotification().parse(sampleNotification.get("bt_signature"), sampleNotification.get("bt_payload"));
+
+        assertEquals(WebhookNotification.Kind.RECIPIENT_UPDATED_GRANTED_PAYMENT_METHOD, notification.getKind());
 
         GrantedPaymentInstrumentUpdate update = notification.getGrantedPaymentInstrumentUpdate();